+++ apps/openmw-mp/Inventory.hpp
@@ -25,7 +25,7 @@ public:
 
     //inventory
     int getChangesSize() const;
-    void addItem(const std::string& refId, unsigned int count, int charge, int enchantmentCharge);
+    void addItem(const std::string& refId, unsigned int count, int charge, double enchantmentCharge);
     void removeItem(const std::string& refId, unsigned short count);
 
     /**
@@ -33,11 +33,11 @@ public:
      * @param slot
      * @return refid, count, charge, enchantmentCharge
      */
-    std::tuple<std::string,int, int, int> getInventoryItem(unsigned int slot) const;
+    std::tuple<std::string,int, int, double> getInventoryItem(unsigned int slot) const;
 
 
     // equipment
-    void equipItem(unsigned short slot, const std::string& refId, unsigned int count, int charge, int enchantmentCharge);
+    void equipItem(unsigned short slot, const std::string& refId, unsigned int count, int charge, double enchantmentCharge);
     void unequipItem(unsigned short slot);
 
     bool hasItemEquipped(const std::string& refId) const;
@@ -47,7 +47,7 @@ public:
      * @param slot
      * @return refid, count, charge, enchantmentCharge
      */
-    std::tuple<std::string,int, int, int> getEquipmentItem(unsigned short slot) const;
+    std::tuple<std::string,int, int, double> getEquipmentItem(unsigned short slot) const;
 
 
 private:
